refactored splashpage design
