+++ src/Components/Login/NavbarLoginForm.css
@@ -36,7 +36,7 @@
 add onHover to give them background colors from our palette
 typography to make the font look better */
 
-.navbar__items {
+.navbar-items {
   padding-right: 1rem;
   margin-top: auto;
 }
@@ -45,18 +45,18 @@ typography to make the font look better */
   padding-right: 1rem;
 }
 
-.navbar__tab__container {
+.navbar__tab-container {
   position: relative;
   text-align: center;
   display: inline-flex;
   padding-right: 2.2rem;
 }
 
-.navbar__tab__text {
+.navbar__tab-text {
     position: absolute;
     top: 60%;
-    left: 20%;
-    transform: translate(-28%, -50%);
+    left: 25%;
+    transform: translate(-28%, -60%);
     color: #2C92A8;
     /* font-size: var(--;font-size-small);
     font-weight: var(--font-weight-bold); */
@@ -64,6 +64,34 @@ typography to make the font look better */
     font-weight: var(--font-weight-extra-bold);
     text-transform: uppercase;
     letter-spacing: .2rem;
+}
+
+.tab-pink:hover {
+  color: #4B1E75;
+}
+
+.tab-blue:hover {
+  color: #237E74
+}
+
+.tab-yellow:hover {
+  color: #F6A04D
+}
+
+.tab-login {
+  color: #4B1E75;
+}
+
+.tab-login:hover {
+  color: #237E74;
+}
+
+.tab-signup {
+  color: #4B1E75;
+}
+
+.tab-signup:hover {
+  color: #E3ED55
 }
 
 .navbar__button {
@@ -74,12 +102,12 @@ typography to make the font look better */
   width: 8rem;
 }
 
-.navbar__tab__image {
-  width: 9rem;
+.navbar__tab-image {
+  width: 10rem;
   opacity: 0;
 }
 
-.navbar__tab__container:hover .navbar__tab__image {
-  width: 8rem;
+.navbar__tab-container:hover .navbar__tab-image {
+  width: 10rem;
   opacity: 1;
 }
